Move common settings to one file
