+++ longclaw/checkout/api.py
@@ -9,10 +9,11 @@ from rest_framework.response import Response
 from longclaw.basket.utils import get_basket_items, destroy_basket
 from longclaw.orders.models import Order, OrderItem, Address
 from longclaw.checkout.models import ShippingCountry
-from longclaw.checkout import app_settings, serializers
+from longclaw.checkout import serializers
 from longclaw.checkout.utils import PaymentError
+from longclaw import settings
 
-gateway = import_string(app_settings.PAYMENT_GATEWAY)()
+gateway = import_string(settings.PAYMENT_GATEWAY)()
 
 @api_view(['GET'])
 @permission_classes([permissions.AllowAny])
