+++ longclaw/checkout/gateways/stripe_payment.py
@@ -1,7 +1,7 @@
 import math
 from django.conf import settings
 import stripe
-from longclaw.checkout.app_settings import CURRENCY
+from longclaw.settings import CURRENCY
 from longclaw.checkout.utils import PaymentError
