--- longclaw/orders/app_settings.py
@@ -1,3 +1,0 @@
-from django.conf import settings
-
-PRODUCT_VARIANT_MODEL = getattr(settings, 'PRODUCT_VARIANT_MODEL', 'products.ProductVariant')
