+++ longclaw/orders/wagtail_hooks.py
@@ -10,8 +10,9 @@ class OrderModelAdmin(ModelAdmin):
     menu_icon = 'list-ul'
     add_to_settings_menu = False
     exclude_from_explorer = False
-    list_display = ('status', 'status_note', 'email', 'payment_date')
+    list_display = ('id', 'status', 'status_note', 'email', 'payment_date', 'total_items', 'total')
     list_filter = ('status', 'payment_date', 'email')
     inspect_view_enabled = True
+    index_template_name = 'orders_index.html'
 
 modeladmin_register(OrderModelAdmin)
