+++ longclaw/products/serializers.py
@@ -1,10 +1,9 @@
-from django.conf import settings
+from longclaw import settings
 from django.apps import apps
 from rest_framework import serializers
 from longclaw.products.models import Product
 
-PRODUCT_VARIANT_MODEL = getattr(settings, 'PRODUCT_VARIANT_MODEL', 'products.ProductVariant')
-ProductVariant = apps.get_model(*PRODUCT_VARIANT_MODEL.split('.'))
+ProductVariant = apps.get_model(*settings.PRODUCT_VARIANT_MODEL.split('.'))
 
 class ProductSerializer(serializers.ModelSerializer):
